+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Unreleased
+
+- Improved unit test coverage.
+- Added cmdlet `Set-CosmosDBUser` for setting the user Id of an
+  existing user.
+
 ## 1.0.1.40
 
 - Minor bugfixes.